fixed issue with search leave applications and adding stateid as search parameter

+++ hr/hr-leave/src/main/java/org/egov/eis/model/LeaveApplication.java
@@ -85,9 +85,9 @@ public class LeaveApplication {
     @JsonFormat(pattern = "dd/MM/yyyy")
     private Date compensatoryForDate;
 
-    private Double leaveDays;
+    private Float leaveDays;
 
-    private Double availableDays;
+    private Float availableDays;
 
     private Integer halfdays;
